requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.124.4
h11==0.16.0
httptools==0.7.1
idna==3.11
makefun==1.16.0
motor==3.7.1
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
